refactor(core): extract maintenance/backfill-narrative.ts — session-summary narrative backfill
